Remove RTreeLookup and move DeterministicLookup to builtins
